refactor simd store and load
